pyproject.toml
[tool.poetry]
name = "rule-engine"
version = "0.1.0"
description = ""
authors = ["Eyal Segal <segaleseo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.1.2"
toml = "^0.10.2"
pyodbc = "^4.0.32"
Flask-SQLAlchemy = "^2.5.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
